move view_to_component()
